% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadData.R
\name{downloadData}
\alias{downloadData}
\title{Import TCGA metadata}
\usage{
downloadData(
  URL = NULL,
  fileName = NULL,
  fileType = "EXCEL",
  directory = file.path(".", "TCGA"),
  sheetIndex = 1,
  verbose = TRUE
)
}
\arguments{
\item{URL}{\code{character} A remote URL for the clinical information}

\item{fileName}{\code{character} A new file name to be saved}

\item{fileType}{\code{character} A file type}

\item{directory}{\code{character} A path to a directory where the file
will be saved, create a directory if it does not exist}

\item{sheetIndex}{\code{numeric} The index of the sheet (.xlsx) to read data from}

\item{verbose}{\code{logical} Print status message during download,
defaults to TRUE}
}
\value{
A downloaded matrix
}
\description{
Download TCGA metadata (.xlsx) from the GDC data portal
}
\examples{
tcga <- downloadData()

}
\author{
Heewon Seo, \email{Heewon.Seo@uhnresearch.ca}
}
