% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticalTesting.R
\name{statisticalTesting}
\alias{statisticalTesting}
\title{Perform statistical testing}
\usage{
statisticalTesting(df, categorial = FALSE)
}
\arguments{
\item{df}{\code{data frame} A data frame that has two columns (column_one
and column_two). For categorical variables, column_one and column_two
possess each category. On the other hand, column_one must be a numeric
vector, while column_two contains the factor to be compared for
continuous variables.}

\item{categorical}{\code{logical} A logical value whether input is categorical variables}
}
\value{
A list of statistical results
}
\description{
Carry out statistical testing depending on variables
(either categorical or discrete variables)
}
\examples{
tcga <- downloadData()

df <- data.frame(column_one = tcga$gender, column_two = tcga$vital_status)
statisticalTesting(df, categorical = TRUE)

df <- data.frame(column_one = tcga$age_at_initial_pathologic_diagnosis,
    column_two = tcga$race)
statisticalTesting(df, categorical = FALSE)

}
\author{
Heewon Seo, \email{Heewon.Seo@uhnresearch.ca}
}
