% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationBox.R
\name{visualizationBox}
\alias{visualizationBox}
\title{Generate a boxplot}
\usage{
visualizationBox(
  object,
  cancer_type = NULL,
  patients = NULL,
  phenotype,
  metadata_col,
  metadata_levels
)
}
\arguments{
\item{object}{\code{matrix} Patient information (matrix)}

\item{cancer_type}{\code{vector} A vector of cancer types of interest from
'type' column}

\item{patients}{\code{vector} A vector of patient barcodes from
'bcr_patient_barcode' column}

\item{phenotype}{\code{character} A string of a specific phenotype (column name)}

\item{metadata_col}{\code{character} A string of the column name (phenotype)
to be split into groups}

\item{metadata_levels}{\code{vector} A vector of two values (factors) to be
compared}
}
\value{
None/plot a box plot
}
\description{
Generate a boxplot to compare a specific phenotype for a set of samples.
}
\examples{
tcga <- downloadData()

visualizationBox(tcga, cancer_type = c("BLCA", "BRCA"),
    phenotype = "age_at_initial_pathologic_diagnosis",
    metadata_col = "race", 
    metadata_levels = c("WHITE", "ASIAN"))

visualizationBox(tcga, cancer_type = c("BLCA", "BRCA"),
    phenotype = "age_at_initial_pathologic_diagnosis",
    metadata_col = "race", 
    metadata_levels = c("WHITE", "ASIAN", "BLACK OR AFRICAN AMERICAN"))

}
\author{
Heewon Seo, \email{Heewon.Seo@uhnresearch.ca}
}
